Customer_Segmentation
The agency uses this project to analyze customer behavior in the motor insurance dataset using Exploratory Data Analysis and clustering. The goal is to extract meaningful insights to assess whether premiums align with actual risk and whether customer segments are effectively identified and targeted to improve pricing, retention, and acquisition strategies.